+++ packages/ragbits-core/src/ragbits/core/vector_stores/pgvector.py
@@ -10,11 +10,12 @@ from pydantic.json import pydantic_encoder
 from ragbits.core.audit import trace
 from ragbits.core.embeddings.base import Embedder
 from ragbits.core.vector_stores.base import (
+    EmbeddingType,
     VectorStoreEntry,
-    VectorStoreNeedingEmbedder,
     VectorStoreOptions,
     VectorStoreOptionsT,
     VectorStoreResult,
+    VectorStoreWithExternalEmbedder,
     WhereQuery,
 )
 
@@ -31,7 +32,7 @@ DISTANCE_OPS = {
 
 
 # TODO: Add support for image embeddings
-class PgVectorStore(VectorStoreNeedingEmbedder[VectorStoreOptions]):
+class PgVectorStore(VectorStoreWithExternalEmbedder[VectorStoreOptions]):
     """
     Vector store implementation using [pgvector]
 
@@ -50,8 +51,6 @@ class PgVectorStore(VectorStoreNeedingEmbedder[VectorStoreOptions]):
         distance_method: str = "cosine",
         hnsw_params: dict | None = None,
         default_options: VectorStoreOptions | None = None,
-        embedding_name_text: str = "text",
-        embedding_name_image: str = "image",
     ) -> None:
         """
         Constructs a new PgVectorStore instance.
@@ -64,15 +63,11 @@ class PgVectorStore(VectorStoreNeedingEmbedder[VectorStoreOptions]):
             distance_method: The distance method to use.
             hnsw_params: The parameters for the HNSW index. If None, the default parameters will be used.
             default_options: The default options for querying the vector store.
-            embedding_name_text: The name of the embedding for text.
-            embedding_name_image: The name of the embedding for images.
         """
         (
             super().__init__(
                 default_options=default_options,
                 embedder=embedder,
-                embedding_name_text=embedding_name_text,
-                embedding_name_image=embedding_name_image,
             ),
         )
 
@@ -236,15 +231,13 @@ class PgVectorStore(VectorStoreNeedingEmbedder[VectorStoreOptions]):
             entries=entries,
             vector_size=self._vector_size,
             embedder=repr(self._embedder),
-            embedder_for_text=self._embedding_name_text,
-            embedder_for_image=self._embedding_name_image,
         ):
             embeddings = await self._create_embeddings(entries)
 
             try:
                 async with self._client.acquire() as conn:
                     for entry in entries:
-                        if entry.id not in embeddings or self._embedding_name_text not in embeddings[entry.id]:
+                        if entry.id not in embeddings or EmbeddingType.TEXT not in embeddings[entry.id]:
                             warnings.warn(f"Skipping entry {entry.id} as it has no text embeddings.")
                             continue
 
@@ -252,7 +245,7 @@ class PgVectorStore(VectorStoreNeedingEmbedder[VectorStoreOptions]):
                             insert_query,
                             str(entry.id),
                             entry.text,
-                            str(embeddings[entry.id][self._embedding_name_text]),
+                            str(embeddings[entry.id][EmbeddingType.TEXT]),
                             json.dumps(entry.metadata, default=pydantic_encoder),
                         )
             except asyncpg.exceptions.UndefinedTableError:
@@ -348,27 +341,24 @@ class PgVectorStore(VectorStoreNeedingEmbedder[VectorStoreOptions]):
             vector_size=self._vector_size,
             distance_method=self._distance_method,
             embedder=repr(self._embedder),
-            embedder_for_text=self._embedding_name_text,
-            embedder_for_image=self._embedding_name_image,
         ) as outputs:
             vector = (await self._embedder.embed_text([text]))[0]
 
+            query_options = (self.default_options | options) if options else self.default_options
             retrieve_query, values = self._create_retrieve_query(vector, query_options)
 
             try:
                 async with self._client.acquire() as conn:
                     results = await conn.fetch(retrieve_query, *values)
 
-                retrieved_entries = [
+                outputs.results = [
                     VectorStoreResult(
                         entry=VectorStoreEntry(
                             id=record["id"],
                             text=record["key"],
                             metadata=json.loads(record["metadata"]),
                         ),
-                        vectors={
-                            self._embedding_name_text: json.loads(record["vector"]),
-                        },
+                        vector=json.loads(record["vector"]),
                         score=record["distance"],
                     )
                     for record in results
@@ -376,9 +366,8 @@ class PgVectorStore(VectorStoreNeedingEmbedder[VectorStoreOptions]):
 
             except asyncpg.exceptions.UndefinedTableError:
                 print(f"Table {self._table_name} does not exist.")
-                retrieved_entries = []
-            outputs.retrieved_entries = retrieved_entries
-            return retrieved_entries
+                outputs.results = []
+            return outputs.results
 
     async def list(
         self, where: WhereQuery | None = None, limit: int | None = None, offset: int = 0